requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.14.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
squarify==0.4.4
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0
torchmetrics==1.7.1
tornado==6.4.2
traitlets==5.14.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
